Add groups folder, move navbar to groups
